requirements.txt
Django==2.2.4
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.11.0
